+++ R/JointAI.R
@@ -169,7 +169,7 @@ utils::globalVariables(c("i", "value", "chain", "iteration"))
 #' @param keep_aux logical; Should constant effects of auxiliary variables be kept in the output?
 #' @param ridge logical; should the parameters of the main model be penalized using ridge regression? Default is \code{FALSE}
 #' @param parallel logical; should the chains be sampled using parallel computation? Default is \code{FALSE}
-#' @param ncores number of cores to use for parallel computation; if left empty all except two cores will be used
+#' @param n.cores number of cores to use for parallel computation; if left empty all except two cores will be used
 #' @param seed optional seed value for reproducibility
 #' @param ppc logical: should monitors for posterior predictive checks be set? (not yet used)
 #' @name sharedParams